refactor: finish renaming SignUpForm folder
